:recycle: Move the logic to the lib folder
