+++ packages/ntm_cli/bin/ntm_cli.dart
@@ -1,8 +1,6 @@
 import 'dart:io';
 
-import 'package:ntm/ntm.dart';
-
-final _ntm = Ntm();
+import 'package:ntm_cli/ntm_cli.dart';
 
 void main(List<String> arguments) {
   exitCode = 0; // Presume success.
@@ -11,43 +9,8 @@ void main(List<String> arguments) {
     exitCode = 2;
     stderr.writeln('Usage: ntm [script]');
   } else if (arguments.length == 1) {
-    _runFile(arguments.first);
-    if (_hadError) {
-      exitCode = 2;
-    }
+    runFile(arguments.first);
   } else {
-    _runPrompt();
+    runPrompt();
   }
-}
-
-void _runFile(String path) {
-  final file = File(path);
-  _run(file.readAsStringSync());
-}
-
-void _runPrompt() {
-  while (true) {
-    stdout.write('[ntm] > ');
-    final line = stdin.readLineSync();
-    if (line == null) {
-      break;
-    }
-    _run(line);
-    _hadError = false;
-  }
-}
-
-void _run(String script) {
-  _ntm.run(script);
-}
-
-var _hadError = false;
-
-void report({
-  required int line,
-  required String where,
-  required String message,
-}) {
-  stderr.writeln('[line $line] Error $where: $message');
-  _hadError = true;
 }
