--- packages/ntm_cli/example/ntm_cli_example.dart
@@ -1,6 +1,0 @@
-import 'package:ntm_cli/ntm_cli.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}
