+++ packages/ntm_cli/lib/ntm_cli.dart
@@ -1,3 +1,3 @@
 library ntm_cli;
 
-export 'src/ntm_cli_base.dart';
+export 'src/ntm_cli.dart';
